Move test folder to correct directory position
