+++ test/resources/junit/failures/TEST-features-secondary-srv_delete_channel_from_ui.xml
