+++ test/resources/junit/failures/TEST-features-secondary-trad_centos_client.xml
