+++ test/resources/junit/passed/TEST-features-secondary-srv_delete_channel_from_ui.xml
