+++ test/resources/junit/passed/TEST-features-secondary-srv_power_management.xml
