+++ test/resources/templates/mail-template-jenkins-env-fail.txt
