+++ test/resources/templates/mail-template-jenkins.txt
