+++ test/test_terraformer.py
@@ -1,28 +1,27 @@
-import terraformer
+from terracumber import terraformer
 import unittest
 from subprocess import CalledProcessError
 from unittest.mock import patch
 
-@patch('terraformer.copy')
-@patch('terraformer.path')
-@patch('terraformer.symlink')
-@patch('terraformer.unlink')
+
+@patch('terracumber.terraformer.copy')
+@patch('terracumber.terraformer.path')
+@patch('terracumber.terraformer.symlink')
+@patch('terracumber.terraformer.unlink')
 class TestGit(unittest.TestCase):
     def setUp(self):
         self.terraform_path = 'test/resources'
         self.maintf = 'test/resources/test.tf'
         self.backend = 'libvirt'
-        self.variables = { 'CUCUMBER_BRANCH': 'test' }
+        self.variables = {'CUCUMBER_BRANCH': 'test'}
         self.output_file = 'test/resources/output.log'
-
 
     def test_manage_backend_symlink(self, mock_unlink, mock_symlink, mock_path, mock_copy):
         mock_path.exists.return_value = True
         mock_path.abspath.return_value = '/tmp'
         self.terraformer = terraformer.Terraformer(self.terraform_path, self.maintf, self.backend)
         mock_unlink.assert_called_once_with('test/resources/modules/backend')
-        mock_symlink.assert_called_once_with('/tmp/backend_modules/libvirt','test/resources/modules/backend')
-
+        mock_symlink.assert_called_once_with('/tmp/backend_modules/libvirt', 'test/resources/modules/backend')
 
     def test_get_hostname(self, mock_unlink, mock_symlink, mock_path, mock_copy):
         # Test an invalid machine
@@ -34,37 +33,36 @@ class TestGit(unittest.TestCase):
         with self.assertRaises(FileNotFoundError):
             self.assertIsNone(self.terraformer.get_hostname('controller'))
 
-
     def test_get_resources(self, mock_unlink, mock_symlink, mock_path, mock_copy):
         self.terraformer = terraformer.Terraformer(self.terraform_path, self.maintf, self.backend)
         with patch.object(self.terraformer, '_Terraformer__run_command') as mock_run_command:
             mock_run_command.return_value = ['module.base.module.base_backend.libvirt_network.additional_network[0]',
-'module.base.module.base_backend.libvirt_volume.volumes["opensuse152o"]',
-'module.proxy.module.proxy.module.host.data.template_file.network_config',
-'module.proxy.module.proxy.module.host.data.template_file.user_data',
-'module.proxy.module.proxy.module.host.libvirt_cloudinit_disk.cloudinit_disk[0]',
-'module.proxy.module.proxy.module.host.libvirt_domain.domain[0]',
-'module.proxy.module.proxy.module.host.libvirt_volume.main_disk[0]',
-'module.proxy.module.proxy.module.host.null_resource.provisioning[0]',
-'module.server.module.server.module.host.data.template_file.network_config',
-'module.server.module.server.module.host.data.template_file.user_data',
-'module.server.module.server.module.host.libvirt_cloudinit_disk.cloudinit_disk[0]',
-'module.server.module.server.module.host.libvirt_domain.domain[0]',
-'module.server.module.server.module.host.libvirt_volume.main_disk[0]',
-'module.server.module.server.module.host.null_resource.provisioning[0]']
+                                             'module.base.module.base_backend.libvirt_volume.volumes["opensuse152o"]',
+                                             'module.proxy.module.proxy.module.host.data.template_file.network_config',
+                                             'module.proxy.module.proxy.module.host.data.template_file.user_data',
+                                             'module.proxy.module.proxy.module.host.libvirt_cloudinit_disk.cloudinit_disk[0]',
+                                             'module.proxy.module.proxy.module.host.libvirt_domain.domain[0]',
+                                             'module.proxy.module.proxy.module.host.libvirt_volume.main_disk[0]',
+                                             'module.proxy.module.proxy.module.host.null_resource.provisioning[0]',
+                                             'module.server.module.server.module.host.data.template_file.network_config',
+                                             'module.server.module.server.module.host.data.template_file.user_data',
+                                             'module.server.module.server.module.host.libvirt_cloudinit_disk.cloudinit_disk[0]',
+                                             'module.server.module.server.module.host.libvirt_domain.domain[0]',
+                                             'module.server.module.server.module.host.libvirt_volume.main_disk[0]',
+                                             'module.server.module.server.module.host.null_resource.provisioning[0]']
             expected_result = ['module.proxy.module.proxy.module.host.libvirt_domain.domain[0]',
-'module.proxy.module.proxy.module.host.libvirt_volume.main_disk[0]',
-'module.server.module.server.module.host.libvirt_domain.domain[0]',
-'module.server.module.server.module.host.libvirt_volume.main_disk[0]']
+                               'module.proxy.module.proxy.module.host.libvirt_volume.main_disk[0]',
+                               'module.server.module.server.module.host.libvirt_domain.domain[0]',
+                               'module.server.module.server.module.host.libvirt_volume.main_disk[0]']
             self.assertEqual(self.terraformer._Terraformer__get_resources('.*(domain|main_disk).*'), expected_result)
             self.assertFalse(self.terraformer._Terraformer__get_resources('.*invalid.*'))
 
-
     @patch('builtins.open')
-    @patch('terraformer.Popen')
+    @patch('terracumber.terraformer.Popen')
     def test_run_command(self, mock_popen, mock_open, mock_unlink, mock_symlink, mock_path, mock_copy):
         mock_popen.return_value.wait.return_value = 0
-        self.terraformer = terraformer.Terraformer(self.terraform_path, self.maintf, self.backend, None, self.output_file)
+        self.terraformer = terraformer.Terraformer(self.terraform_path, self.maintf, self.backend, None,
+                                                   self.output_file)
         with patch.object(self.terraformer, '_Terraformer__run_command_iterator') as mock_cmd_iterator:
             # Test log output
             mock_cmd_iterator.return_value = iter(["TEST"])
@@ -88,8 +86,7 @@ class TestGit(unittest.TestCase):
             mock_open.return_value.write.assert_not_called()
             mock_open.return_value.close.assert_not_called()
 
-
-    @patch('terraformer.Popen')
+    @patch('terracumber.terraformer.Popen')
     def test_run_command_iterator(self, mock_popen, mock_unlink, mock_symlink, mock_path, mock_copy):
         self.terraformer = terraformer.Terraformer(self.terraform_path, self.maintf, self.backend, {}, self.output_file)
         iterator = self.terraformer._Terraformer__run_command_iterator(mock_popen)
@@ -98,6 +95,18 @@ class TestGit(unittest.TestCase):
             self.assertEqual(next(iterator), "TEST")
             mock_popen.stdout.readline.return_value = ""
             next(iterator)
+
+    def test_apply(self, mock_unlink, mock_symlink, mock_path, mock_copy):
+        # Arrange
+        self.terraformer = terraformer.Terraformer(self.terraform_path, self.maintf, self.backend, None,
+                                                   self.output_file)
+        with patch.object(self.terraformer, '_Terraformer__run_command') as mock_run_command:
+
+            # Act
+            self.terraformer.apply(20)
+
+            # Assert
+            mock_run_command.assert_called_with(["/usr/bin/terraform", "apply", "-auto-approve", "-parallelism=20"])
 
 
 if __name__ == '__main__':
